action: fix permission errors on vanilla GHA runners
There were various permission errors when running this action on standard GHA runners
because they do not run as root by default. We need to update the action in a few places
to use sudo where necessary and update the permissions on the image to make it properly
readable and writable in order to fix this.

Signed-off-by: William Findlay <will@isovalent.com>